Rename to match interface name
