+++ resources/assets/lib/interfaces/score-json.ts
